extract the conditional #authenticated? logic into a module
